The variable selection mechanism for what goes into the CLM DART vector can now come from 3 sources ... the restart file, and two additional history files. It is now demonstrated how to create a history file in the 'vector' format used by the restart files. This should allow for more accurate forward observation operators. There is also now a namelist-driven range-restriction capability for the variables that will be updated in the CLM restart file. It is also now possible to make a run-time specification of whether or not you want to update the variable in the restart file.
In order to remove confusion about this, a namelist variable had to change from
'clm_state_variables' to simply 'clm_variables' because some of the variables
in the DART vector are not technically 'state' in the strictest sense of the word.

The forward observation operators for brightness temperature may now explicitly grab
what they need from the DART state via the model_interpolate() function as opposed to
having to open,read,close a file (for each ensemble member) for every single observation.
Furthermore, the data can be at the same resolution as the restart file. All good.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@7195 acad1b26-7d42-44b9-9cec-aac8293558f5