Committing the last changes to fix the kind/type confusion in get_close_obs().
Also added the statements to get_close_obs to save the resulting vertically 
converted obs back in the incoming array so that next time this location is 
used we won't have to convert it again.

Removed two unused variables. 

Removed the local redefinition of a module variable for the random number sequences.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@7340 acad1b26-7d42-44b9-9cec-aac8293558f5