These scripts handle the FV case by default. The input.nml has a separate (commented out) block for SE configurations. The perfect_model.csh, assimilate.csh and CESM_DART_config can handle FV, SE, and WACCM configurations with no modification.
The instructions how to modify the CESM*setup* scripts for SE and WACCM
configurations will be put in the model_mod.html to keep them from 
invading already-complicated 'simple' versions of the setup scripts.

There is a new SourceMods tarfile 
http://www.image.ucar.edu/pub/DART/CESM/DART_SourceMods_cesm1_1_1_19Mar2015.tar
that has a modified lt_archive.sh as well as a src.cam/dynamics/fv/cd_core.F90
that helps stabilize WACCM runs but is not recommended to be used for non-WACCM 
cases. The scripts automatically install or remove cd_core.F90 as needed.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@7742 acad1b26-7d42-44b9-9cec-aac8293558f5