Mostly formatting tweaks in the html files to make sure the 'top' link is inline with the horizontal rule. Real important - not.
Added some bulletproofing to the matlab scripts to prevent Matlab
from making 1D (character) arrays into column-major when character 
arrays should be row-major.

Added some useful options to the gfortran compiler template.

Added support for an upper atmosphere observation that allows the
SSUSI observation converter to compile (sheesh!) the converter is already
in the trunk and is in use on the tiegcm branch.

The changes to assim_model_mod.f90 are improved commenting - nothing executable.
obs_utilities_mod.f90 has a new routine getvarshape() to return the array 
of sizes for any netCDF variable. 

Added tutorial/index.html and listed it in the DART/index.html



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@7959 acad1b26-7d42-44b9-9cec-aac8293558f5