This is expected to be the last commit of the old code style model_mod.f90. The next version will have extensive style changes, but bit-for-bit agreement with this one.
It has important fixes for errors in model_heights, etc., which were introduced
in the upgrade from using the nearest grid point for calculating heights on model
levels to interpolating to the ob location.  There was an extra division of
the surface geopotential by grav.  There was also an error in the conversion from
a 'do while' loop to a 'do ... ; if' structure when searching for the bounding
heights of an ob at a height.
These had no effect on assimilations of obs which were not on heights,
and at worst resulted in obs on heights being ignored when they shouldn't have been.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@9470 acad1b26-7d42-44b9-9cec-aac8293558f5