This function creates a trellis-like plot of the vertical summary statistics for two obs_diag_output.nc files.
At present, it uses a hardwired number of rows and columns for the trellis,
this should be user input.

At present, ALL candidate observation types are plotted.
One figure has the change in RMSE.
One figure has the change in abs(BIAS).
One figure has the change in number of observations used.

One figure has the PERCENT of observations used for both experiments
and the X axis has the maximum number of observations for that region.
This is the tricky one ... the first X ticklabel is the maximum observation
count for region 1, the second label is for region 2, etc. The way the
scaling takes place, the first ticklabel appears to be the 
baseline for the SECOND region. This is NOT the case!

There's a lot more work that could have been done on this, you won't hurt
my feelings suggesting improvements.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@9580 acad1b26-7d42-44b9-9cec-aac8293558f5