fix a bug where the obs error field was too small to contain values > 10.  one of our users has a file from 2015 which has wind errors of > 10 for satellite winds.
changing the write format in the 2 prepbufr converters
requires a corresponding change in the read format in
real_obs_mod.

the "real" fix is to stop writing ascii intermediate files
and combine these two converters into a single program.
at the end of the prepbufr code we have all the bits we
need to create a dart observation, except the date.  (the
files contain hours only; the date comes from the filename.)

for now this allows users to convert these observations
without error.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@9759 acad1b26-7d42-44b9-9cec-aac8293558f5