a collection of minor code changes to produce more correct and hopefully more informative error messages when a dart restart file is unable to be opened.  specific changes are:
assim_model_mod.f90:  update to open_restart_read() for better
error messages ('unable to find file' message is different from
'unable to open file' message is different from 'unable to read 
a valid timestamp from the file').  now calls the open_file() and
close_file() routines from the utility mod.

utilities_mod.f90:  added an optional last argument to the
open_file() function to allow the read error status value to
be passed back to the caller for further processing or to add
more context to the error messages.  removed the code block
in this function which has been commented out since the first
CVS commit, revision 59, in 2002.

time_manager_mod.f90:  read_time() now errors out not just when
the read() call has a non-zero status but also when the numeric
time values are illegal.  since these are being read from a file
there are always 2 numbers so they must both be positive and
the seconds must be within range. this test can help identify
files which are corrupted or written on an architecture with
a different byte order. commented out the call to the 
verbose dump_attributes() routine since the filename and
the read status code will be printed and should identify
the problem more succinctly.  in set_time() moved the lines
that format up two error messages to inside the if() blocks 
so they aren't executed until an error actually occurs.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@10339 acad1b26-7d42-44b9-9cec-aac8293558f5