Correctly setting the QC values for the (computed) *HORIZONTAL_WIND metrics. Previous version were not honoring the incoming qc value and so the number of possible observations was getting over-reported.
This version is a little closer to the style guide requirements
and reinstates some of the indenting that got hosed with commit 10348.

This version also can be used with obs_seq.in and obs_seq.out files
in a limited fashion.  The number of observations in each region, the 
observation value for the region, the incoming QC for the obs in the regions,
and the summary of what types are actually in the file are all useful -
even if you do not have the prior/posterior quantities.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@10352 acad1b26-7d42-44b9-9cec-aac8293558f5