Bringing up-to-date with respect to the trunk:
Correctly setting the QC values for the (computed) *HORIZONTAL_WIND metrics.
Previous version were not honoring the incoming qc value and so the 
number of possible observations was getting over-reported.

This version is a little closer to the style guide requirements
and reinstates some of the indenting that got hosed with commit 10348.

This version also can be used with obs_seq.in and obs_seq.out files
in a limited fashion.  The number of observations in each region, the 
observation value for the region, the incoming QC for the obs in the regions,
and the summary of what types are actually in the file are all useful -
even if you do not have the prior/posterior quantities.

Removed some debug code that was causing problems (i.e. a fatal error) when people 
input the _edges arrays as opposed to the midpoints to define the vertical binning.
When obs_diag supported the use of the _edges, I had code in to check to make sure 
the new algorithm generated the correct answer under all circumstances.  The old 
algorithm did not work when the user specified the _edges. However, the test that 
was inadvertently left in was to check to make sure the old level index matched 
the new level index (a good thing if you specify the midpoints, 
a bad thing if you specify the edges).

The new algorithm has been working for years now, time to remove the debug code.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@10353 acad1b26-7d42-44b9-9cec-aac8293558f5