commit that goes with rev 10339 with updates to assim_model and utilities_mod to make the log messages when unable to open a dart restart file more helpful.  (this was unintentionally committed to a branch and not to the trunk in the original commit.)  here's the relevant part of the original log msg:
time_manager_mod.f90:  read_time() now errors out not just when
the read() call has a non-zero status but also when the numeric
time values are illegal.  since these are being read from a file
there are always 2 numbers so they must both be positive and
the seconds must be within range. this test can help identify
files which are corrupted or written on an architecture with
a different byte order. commented out the call to the 
verbose dump_attributes() routine since the filename and
the read status code will be printed and should identify
the problem more succinctly.  in set_time() moved the lines
that format up two error messages to inside the if() blocks 
so they aren't executed until an error actually occurs.



git-svn-id: https://subversion.ucar.edu/DAReS/DART/trunk@10365 acad1b26-7d42-44b9-9cec-aac8293558f5